Encapsulate some utils functions inside ImageCaptionEngine as they have not much sense as public API.
